Enable writing ops.expr(a + b * c)
Still need to replace NodeId with Expr<usize>, probably name it ExprId which is a re-usable expression that is part of the computation graph. Expr<T> when T != ExprId is not part of the computation graph and can not be re-used.

So Expr<T> should only implement copy when T implements copy, and Add/Mul/Pow shouldn't implement copy because if they do, it would be easy to add duplicate computations.